Django form displayed in Angular.
By converting a Django model instance to a dict, in Angular the dict is
extracted and a reactive form is created with fields. Problem is
absolutely no validation occurs in Angular and all fields are text. The
problem is if a form is passed from Django to Angular, it cannot be
displayed directly as Angular escapes the html tags for security. But
the body of the form is available and can be extracted manually.