Separated settings and completed production.
In order to achieve production level deployment, CSRF tokens needed to
be implemented. The CSRF token needs to be generated in a Django view by
get_token() and then extracted in the Angular frontend by CookieService.
While performing any POST, PUT or DELETE request, the X-CSRFToken needs
to be set to the CSRF token extracted from the GET request. The
settings.py file has been separated into production and development.
Production is the default and development can be chosen with
--setings=django_angular.settings.development