Checking the CSRF token from get_token.
There seems to be more to do with the CSRF token. The CSRF Middleware
was initially commented out in settings so that is probably why CSRF
token was not being generated. When brought back in, the CSRF token is
always generated even after commenting again. So it seems like browser
stores it for a while. Will need to clear the token in the browser and
check again. It kind of makes sense because the token has an expiry date
and time.