Added content type in POST header.
By doing so, it is possible to directly use DRF API in the backend.
Previously, I had been sending POST data as string which is why I needed
to use json.loads() on the backend. Essentially, looks like, front-end
and back-end have to exchange JSON data only to make this as smooth as
possible.