Ang service and Django serializers for all methods
The company http services moved into a separate service with the company
component merely maintaining the list of the companies and form
handling. For all methods in the Angular http, the Content-Type is
application/json which makes it easier to extract the form contents in
Django and assign it to a serializer object.