Added user registration with password hashing
Used Djangos in-built user model to create a user and hash the password.
Next to use JWT to check for user being logged in. Another issue is CSRF
token. Currently only generated in the new company get request in
Django. But a user could access any link and if there is no CSRF token
in the browser, it may not work. Check it out in detail.