Logged in user with success and fail(401)
Since it is Angular in the frontend, the user object does not get added
to the request like with complete Django using DTL. So the next step
will be to use JWT for checking user is logged in. It is worth trying if
the login_required decorator will still work.