Added user model but can't extract JWT header
Got the basic JWT concept going. Created another table for user tokens
and the JWT is based on the user id and the user name. The problem is
that the JWT token appears in the header in Postman but not in Angular.
This led me to the previous problem of CSRF token. Now I find that even
if CSRF tokens are disabled, POST, DELETE and MODIFY still work.
Removing get_token and clearing the CSRF token in the Chrome removes
that. In Postman, the cookies have to be manually deleted or they
continue to appear in the request and response.