Getting JWT in Angular and fixed CSRF
For the JWT to work, the token needs to be exposed in
CORS_EXPOSE_HEADERS. The name of the header exposed is not in the same
case as that being assigned and same for the way it is received. Need to
check out header conventions. For CSRF, DRF disables CSRF unless the
person is logged in through the DRF sessions and the user object is in
request. So, there is the need to enable csrf_protect for the methods
that need CSRF protection and call according to whether the server is in
dev mode or prod mode.