Refactored CSRF cookie service.
Trying to add a home page which will also pass back in the response a CSRF
token. The problem is get_token does not generate tokens. It extracts a
CSRF token from the request header. To set a CSRF token, there seems to
be another method. A request to the company list GET does produce a CSRF
token. The question is where and how is this being produced? Is it
automatic? If so, why does it not happen in the home page?