Added user to company mode, in POST/GET methods
Had issues with adding the user as a foreign key because then adding it
to the company serializer gave the error that the user object cannot be
serializer. This is because the user was embedded inside the company. So
added the user id only to the company serializer. Basic company
creation and fetch works. More work needed on these methods also.