Angular error messages on modify, delete and fetch
The only problem is the placement of the errors. If a form is being
submitted, the error is displayed below the form. But if a delete error
exists, the error is shown at the top of the list. to show it where it
happens, will need a better error handling mechanism.