Rearranged static files and Django url.py
Used the approach of loading the Angular app if any request other than
an API request comes along. To serve the Angular app, used TemplateView
with template name rather than serve which seems deprecated. Also,
loading the JS files as static files rather than doing a redirect in
urls.py. To build, use --output-hashing none to remove the hashes
appended to every build. The --prod does not work, there is an error
that needs to be debugged.